Cluster/spark_apps/requirements.txt
tensorflow
tweet-preprocessor
nltk
pyspark
wordcloud
pandas